chore: separate components
